Split MatrixTranform classes into separate files, adjust Light class
