+++ geSG/src/geSG/Light.h
@@ -7,26 +7,19 @@ namespace ge
    namespace sg
    {
       /**
-       * Holds information about the point or directional light. The usage is up to the user.
-       * You can put it into a scene graph and make a visualisation technique that will lit only
-       * the scene beyond the light node. You can make VT that takes light and uses it as global
-       * sun light or headlight.
+       * Holds information about point lights and directional lights, serves as a base class for spotlights.
        */
       struct Light
       {
-         glm::vec4 position; ///< Position in homogenous coords. Set w to 0 to have directional light. The semantic is position so (1,0,0,0) means light shines at direction (-1,0,0,0).
-         glm::vec4 color; ///< light intensity in RGBA (the A could be used as user sees fit)
+          glm::vec4 position; ///< position in homogenous coords, if w == 0, it is a directional light and the first three components contain its direction
+          glm::vec4 color; ///< light intensity in RGBA (the A could be used as user sees fit)
       };
 
-      struct DirectionalLight : public Light
+      struct SpotLight : public Light
       {
-          glm::vec3 direction; ///< direction of the center of the directional light or spot light
-      };
-
-      struct SpotLight : public DirectionalLight
-      {
-         float innerAngle; ///< inner angle (hot spot) in radians
-         float outerAngle; ///< outer angle (falloff) in radians
+          glm::vec3 direction; ///< vector containing the direction of the center of the spotlight cone 
+          float innerAngle; ///< inner angle (hot spot) in radians
+          float outerAngle; ///< outer angle (falloff) in radians
       };
    }
 }
